requirements.txt
streamlit
yt-dlp
openai-whisper
torch
torchaudio
numpy
tqdm
speechbrain
scipy
matplotlib
